+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 * [HexaPDF::Type::OutputIntent] for defining output intents
+* [HexaPDF::Document::Metadata#delete] for deleting metadata properties
 * PDF/A metadata properties definitions
 * Added a /Name entry to the default optional content configuration dictionary
   (needed by PDF/A)